Separated "import data" button into "Import data into table" and "Use selection" buttons. Selection of uitable panel works well now.
Tried to add "warning on backtrace" to startup function, but it seems that the line a warning appears won't show up with that command in app designer.

Loading the UserSPDs.xlsx takes 15-25 seconds, which is way longer than I want. I don't understand what makes it so laggy.

Perhaps a couple edit boxes to type in start and end wavelength would be worthwhile, and I could append 0 to them all. I also need to check if any values are <0. Those are small things. I need to package this first.