The program now uses the raw spds imported. It works, and the component channels plot as well.
I need to implement a "lever" that lets us scale the spds, because right now the program assumes that the channels are at 100% output.

I also should verify that the results work as well as before. If not, then I can pre-scale the spds before I go into the runOptimization function and then undo the scaling.