created a function to create a searchpath for any metric name of the spd nested scalar by making use of the tags of each node when we created the tree.
the target paths are made near the beginning of the run optimziation button pushed for the optimization goal metric and stored in myUiFun scalar structure.
The myUiCon structure array, where each entry in the array is a metric, stores the search path in it in the same button pushed for optimization callback. This storing of the search path lets us avoid passing the entire uitable into the fmincon function usage.