Everything works except for exporting the data. I did a lot here, and it was a huge mess of subfunctions.
Note that I also sort place the optimized result and constraint values at the top of the uitable results!

Notably, I split the overloaded property myBestOptimResult into a myBestSpdMix structure that contains all the metrics.

I created a few really nice helper functions that parse through the structure, and I love the idea of using the tags of the uitree constraints for finding a target path (using a list of fieldnames) to take a value from the spd nested structure