Mostly finalized reorganizing.
Each folder in App_Methods has a folder for each UI tab. Files used by a callback in that tab exist in that folder.

For example, The selection tab has a file to prompt and read the spreadsheet, and then the button callback to use the selection of spd powers calls other files and even plots to the selection tab.