Changed dummystruct at startup from ones(401,1) to a gaussian emission with peak 555nm and fwhm of 50 so metrics that use max or min returned a unique value.
Why? The uitree is made via calling the channelPercentsToSPDNestedStruct with a dummy struct, and a new fake function for the example to add new metrics was to show the wavelength of the peak, and becuase it was all 1, it returned an array! Recall that non-scalar metrics are trimmed, so the metric did not appear.

sig = 50/2.355; %fwhm of 50nm to sigma for gauss distr.
dummySpd(:,1) = exp( -(app.wlVecProgram - 555).^2/(2*sig^2));